app/pyproject.toml
[tool.poetry]
name = "tasktest-hse"
version = "0.1.0"
description = ""
authors = ["ilmir <valitovilmir29@gmail.com>"]
readme = "README.md"
packages = [{include = "tasktest_hse"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.3"
django-filter = "^22.1"
gunicorn = "^20.1.0"
django-debug-toolbar = "^3.7.0"
black = "^22.10.0"
isort = "^5.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
